Extract oauth module
