+++ db.json
@@ -1,1 +1,1 @@
-{"_default": {}, "rent2": {}, "rent": {"1": {"note": "Double room 1", "phone": "19496331541", "amount": -976.29}, "3": {"note": "Closet room", "phone": "15308591050", "amount": -876.29}, "2": {"note": "Double room 2", "phone": "14157204626", "amount": -976.28}, "5": {"note": "Big room", "phone": "16178271991", "amount": -1961.28}, "4": {"note": "Window room prorated", "phone": "15308590976", "amount": -1716.28}}, "auth": {"1": {"access_token": "80c9586dd7491171f9648b0d41b46b47e34d70b5e868847350860de47656a2e2"}}, "users": {"1": {"phone": "16178271991", "name": "Nick Lippis"}, "3": {"phone": "15308590976", "name": "Cynthia Laiacona"}, "2": {"phone": "15308591050", "name": "Zack Morris"}, "5": {"phone": "19496331541", "name": "Rob Garbanati"}, "4": {"phone": "14157204626", "name": "Kyle Merwin"}, "7": {"phone": "14153422879", "name": "Ben Hunter"}, "6": {"phone": "18186319598", "name": "Nate Siswanto"}}}
+{"_default": {}, "rent2": {}, "rent": {"1": {"note": "Double room 1", "phone": "19496331541", "amount": -976.29}, "3": {"note": "Closet room", "phone": "15308591050", "amount": -876.29}, "2": {"note": "Double room 2", "phone": "14157204626", "amount": -976.28}, "5": {"note": "Big room", "phone": "16178271991", "amount": -1961.28}, "4": {"note": "Window room prorated", "phone": "15308590976", "amount": -1716.28}}, "auth": {"1": {"access_token": "44ad9b7b56437195d63f67e6819617c4d6fbf3baf8eb47cc0943c4e807754e40"}}, "users": {"1": {"phone": "16178271991", "name": "Nick Lippis"}, "3": {"phone": "15308590976", "name": "Cynthia Laiacona"}, "2": {"phone": "15308591050", "name": "Zack Morris"}, "5": {"phone": "19496331541", "name": "Rob Garbanati"}, "4": {"phone": "14157204626", "name": "Kyle Merwin"}, "7": {"phone": "14153422879", "name": "Ben Hunter"}, "6": {"phone": "18186319598", "name": "Nate Siswanto"}}}
