+++ venmo/settings.py
@@ -1,0 +1,4 @@
+CLIENT_ID = '2667'
+CLIENT_SECRET = 'srDrmU3yf452HuFF63HqHEt25pa5DexZ'
+BASE_URL = "https://api.venmo.com/v1"
+PAYMENTS_BASE_URL = "{base_url}/payments".format(base_url=BASE_URL)
